+++ lib/rubocop/cop/style/method_call_with_args_parentheses.rb
@@ -237,6 +237,7 @@ module RuboCop
           node.parent &&
             (node.parent.pair_type? ||
              node.parent.array_type? ||
+             splat?(node.parent) ||
              ternary_if?(node.parent))
         end
 
@@ -248,7 +249,7 @@ module RuboCop
 
         def call_with_ambiguous_arguments?(node)
           call_with_braced_block?(node) ||
-            call_as_argument?(node) ||
+            call_as_argument_or_chain?(node) ||
             hash_literal_in_arguments?(node) ||
             node.descendants.any? do |n|
               splat?(n) || ternary_if?(n) || logical_operator?(n)
@@ -259,8 +260,8 @@ module RuboCop
           node.block_node && node.block_node.braces?
         end
 
-        def call_as_argument?(node)
-          node.parent && node.parent.send_type?
+        def call_as_argument_or_chain?(node)
+          node.parent && (node.parent.send_type? || node.parent.csend_type?)
         end
 
         def hash_literal_in_arguments?(node)
@@ -275,7 +276,7 @@ module RuboCop
         end
 
         def allowed_chained_call_with_parentheses?(node)
-          return unless cop_config['AllowParenthesesInChaining']
+          return false unless cop_config['AllowParenthesesInChaining']
 
           previous = node.descendants.first
           return false unless previous && previous.send_type?