+++ spec/rubocop/cop/style/trailing_comma_in_arguments_spec.rb
@@ -23,6 +23,14 @@ RSpec.describe RuboCop::Cop::Style::TrailingCommaInArguments, :config do
       expect_no_offenses('some_method(a, b, c)')
     end
 
+    it 'accepts method call without trailing comma ' \
+       'when a line break before a method call' do
+      expect_no_offenses(<<-RUBY.strip_indent)
+        obj
+          .do_something(:foo, :bar)
+      RUBY
+    end
+
     it 'accepts method call without trailing comma with single element hash' \
         ' parameters at the end' do
       expect_no_offenses('some_method(a: 1)')
@@ -301,6 +309,16 @@ RSpec.describe RuboCop::Cop::Style::TrailingCommaInArguments, :config do
                         a: 0,
                         b: 1,
                      )
+        RUBY
+      end
+
+      it 'accepts a trailing comma in a method call with ' \
+         'a single hash parameter to a receiver object' do
+        expect_no_offenses(<<-RUBY.strip_indent)
+          obj.some_method(
+                            a: 0,
+                            b: 1,
+                         )
         RUBY
       end